Разделил код по модулям
